mobile: Move mocks from library to test (#32308)
