+++ mobile/bazel/kotlin_test.bzl
@@ -3,7 +3,7 @@ load("@io_bazel_rules_kotlin//kotlin:android.bzl", "kt_android_local_test")
 load("@io_bazel_rules_kotlin//kotlin:jvm.bzl", "kt_jvm_test")
 load("//bazel:kotlin_lib.bzl", "native_lib_name")
 
-def _internal_kt_test(name, srcs, deps = [], data = [], jvm_flags = [], repository = "", exec_properties = {}):
+def _internal_kt_test(name, srcs, deps = [], data = [], jvm_flags = [], repository = "", exec_properties = {}, associates = []):
     # This is to work around the issue where we have specific implementation functionality which
     # we want to avoid consumers to use but we want to unit test
     dep_srcs = []
@@ -28,6 +28,7 @@ def _internal_kt_test(name, srcs, deps = [], data = [], jvm_flags = [], reposito
         data = data,
         jvm_flags = jvm_flags,
         exec_properties = exec_properties,
+        associates = associates,
     )
 
 # A simple macro to define the JVM flags that are common for envoy_mobile_jni_kt_test and
@@ -47,7 +48,7 @@ def jvm_flags(lib_name):
 
 # A basic macro to make it easier to declare and run kotlin tests which depend on a JNI lib
 # This will create the native .so binary (for linux) and a .jnilib (for macOS) look up
-def envoy_mobile_jni_kt_test(name, srcs, native_lib_name, native_deps = [], deps = [], repository = "", exec_properties = {}):
+def envoy_mobile_jni_kt_test(name, srcs, native_lib_name, native_deps = [], deps = [], repository = "", exec_properties = {}, associates = []):
     _internal_kt_test(
         name,
         srcs,
@@ -56,6 +57,7 @@ def envoy_mobile_jni_kt_test(name, srcs, native_lib_name, native_deps = [], deps
         jvm_flags = jvm_flags(native_lib_name),
         repository = repository,
         exec_properties = exec_properties,
+        associates = associates,
     )
 
 # A basic macro to make it easier to declare and run kotlin tests
@@ -74,11 +76,11 @@ def envoy_mobile_jni_kt_test(name, srcs, native_lib_name, native_deps = [], deps
 #         "ExampleTest.kt",
 #     ],
 # )
-def envoy_mobile_kt_test(name, srcs, deps = [], repository = "", exec_properties = {}):
-    _internal_kt_test(name, srcs, deps, repository = repository, exec_properties = exec_properties)
+def envoy_mobile_kt_test(name, srcs, deps = [], repository = "", exec_properties = {}, associates = []):
+    _internal_kt_test(name, srcs, deps, repository = repository, exec_properties = exec_properties, associates = associates)
 
 # A basic macro to run android based (robolectric) tests with native dependencies
-def envoy_mobile_android_test(name, srcs, native_lib_name, deps = [], native_deps = [], repository = "", exec_properties = {}):
+def envoy_mobile_android_test(name, srcs, native_lib_name, deps = [], native_deps = [], repository = "", exec_properties = {}, associates = []):
     android_library(
         name = name + "_test_lib",
         custom_package = "io.envoyproxy.envoymobile.test",
@@ -116,4 +118,5 @@ def envoy_mobile_android_test(name, srcs, native_lib_name, deps = [], native_dep
         test_class = "io.envoyproxy.envoymobile.bazel.EnvoyMobileTestSuite",
         jvm_flags = jvm_flags(native_lib_name),
         exec_properties = exec_properties,
+        associates = associates,
     )
